OCT-111: moved GetChannelQuery
