+++ components/catalogs/back/tests/Integration/Infrastructure/Validation/ProductSelection/SystemCriterion/AbstractSystemCriterionTest.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Validation\ProductSelection\SystemCriterion;
 
+use Akeneo\Catalogs\Application\Persistence\Channel\GetChannelQueryInterface;
 use Akeneo\Catalogs\Application\Persistence\Family\GetFamiliesByCodeQueryInterface;
 use Akeneo\Catalogs\Application\Persistence\GetCategoriesByCodeQueryInterface;
-use Akeneo\Catalogs\Application\Persistence\GetChannelQueryInterface;
 use Akeneo\Catalogs\Application\Persistence\Locale\GetChannelLocalesQueryInterface;
+use Akeneo\Catalogs\Infrastructure\Persistence\Channel\GetChannelQuery;
 use Akeneo\Catalogs\Infrastructure\Persistence\Family\GetFamiliesByCodeQuery;
 use Akeneo\Catalogs\Infrastructure\Persistence\GetCategoriesByCodeQuery;
-use Akeneo\Catalogs\Infrastructure\Persistence\GetChannelQuery;
 use Akeneo\Catalogs\Infrastructure\Persistence\Locale\GetChannelLocalesQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
